requirements.txt
# Web UI
streamlit>=1.35.0

# Crawling
requests>=2.31.0
beautifulsoup4>=4.12.0

# Config
python-dotenv>=1.0.0

# RAG + LLM
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
